requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
blinker==1.7.0; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
distro==1.9.0; python_version >= '3.6'
flask==3.0.2; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
openai==1.16.2; python_full_version >= '3.7.1'
pydantic==2.6.4; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
tenacity==8.2.3; python_version >= '3.7'
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
werkzeug==3.0.2; python_version >= '3.8'
youtube-transcript-api==0.6.2
